🐛 Fix : 오류수정
