+++ src/components/study-room/management/StudyManagementPage.tsx
@@ -4,14 +4,41 @@ import StudyExtend from '@/components/study-room/management/StudyExtend';
 import StudyMamagementSkeleton from '@/components/study-room/management/StudyMamagementSkeleton';
 import StudyMemberList from '@/components/study-room/management/StudyMemberList';
 import StudyRoomInfo from '@/components/study-room/management/StudyRoomInfo';
-import { useStudyStore } from '@/stores/studyInfoStore';
+import { getStudyInfoData } from '@/lib/api/study.api';
+
+import { userAuthStore } from '@/stores/userStore';
 import { useParams } from 'next/navigation';
+import { useEffect, useState } from 'react';
 
 export default function StudyManagementPage() {
   const params = useParams();
   const studyId = Number(params.id);
-  const studyInfo = useStudyStore().study;
-  const isLoading = useStudyStore().isLoading;
+  const user = userAuthStore().user;
+
+  const [studyInfo, setStudyInfo] = useState<StudyInfoType>();
+  const [isLoading, setIsLoading] = useState(true);
+
+  useEffect(() => {
+    if (!user) return;
+    // 스터디 정보 가져오기
+    const fetchStudyInfo = async () => {
+      setIsLoading(true);
+      try {
+        const { data } = await getStudyInfoData(studyId);
+        setStudyInfo(data);
+      } catch (error) {
+        console.error('스터디 정보를 불러오지 못했습니다', error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+    fetchStudyInfo();
+  }, [studyId, user]);
+
+  // 스터디 정보 수정
+  const handleUpdate = (newData: StudyInfoUpdateType) => {
+    setStudyInfo((prev) => prev && { ...prev, ...newData });
+  };
 
   return (
     <>
@@ -21,7 +48,11 @@ export default function StudyManagementPage() {
         studyInfo && (
           <>
             <div className='mb-24'>
-              <StudyRoomInfo studyInfoData={studyInfo} studyId={studyId} />
+              <StudyRoomInfo
+                studyInfoData={studyInfo}
+                studyId={studyId}
+                handleUpdate={handleUpdate}
+              />
             </div>
             <div className='mb-24'>
               <StudyMemberList
